+++ sdk/js/src/relayer/relayer/helpers.ts
@@ -6,7 +6,9 @@ import {
   Network,
   tryNativeToHexString,
   isChain,
-  CONTRACTS
+  CONTRACTS,
+  getSignedVAAWithRetry,
+  SignedVaa,
 } from "../../";
 import { BigNumber, ContractReceipt, ethers } from "ethers";
 import { getWormholeRelayer, RPCS_BY_CHAIN } from "../consts";
@@ -20,11 +22,17 @@ import {
   RefundStatus,
   VaaKey,
   DeliveryOverrideArgs,
-  parseForwardFailureError
+  parseForwardFailureError,
 } from "../structs";
-import { DeliveryProvider, DeliveryProvider__factory, Implementation__factory, IWormholeRelayerDelivery__factory } from "../../ethers-contracts/";
-import {DeliveryEvent} from "../../ethers-contracts/WormholeRelayer"
+import {
+  DeliveryProvider,
+  DeliveryProvider__factory,
+  Implementation__factory,
+  IWormholeRelayerDelivery__factory,
+} from "../../ethers-contracts/";
+import { DeliveryEvent } from "../../ethers-contracts/WormholeRelayer";
 import { VaaKeyStruct } from "../../ethers-contracts/IWormholeRelayer.sol/IWormholeRelayer";
+import { NodeHttpTransport } from "@improbable-eng/grpc-web-node-http-transport";
 
 export type DeliveryTargetInfo = {
   status: DeliveryStatus | string;
@@ -37,6 +45,31 @@ export type DeliveryTargetInfo = {
   revertString?: string; // Only defined if status is RECEIVER_FAILURE or FORWARD_REQUEST_FAILURE
   overrides?: DeliveryOverrideArgs;
 };
+
+export async function getVAA(
+  wormholeRPCs: string[] | string,
+  vaaKey: VaaKey,
+  isNode?: boolean
+): Promise<SignedVaa> {
+  if (typeof wormholeRPCs === "string") {
+    wormholeRPCs = [wormholeRPCs];
+  }
+  const vaa = await getSignedVAAWithRetry(
+    wormholeRPCs,
+    vaaKey.chainId! as ChainId,
+    vaaKey.emitterAddress!.toString("hex"),
+    vaaKey.sequence!.toBigInt().toString(),
+    isNode
+      ? {
+          transport: NodeHttpTransport(),
+        }
+      : {},
+    2000,
+    4
+  );
+
+  return Buffer.from(vaa.vaaBytes);
+}
 
 export function parseWormholeLog(log: ethers.providers.Log): {
   type: RelayerPayloadId;
@@ -57,20 +90,25 @@ export function parseWormholeLog(log: ethers.providers.Log): {
 }
 
 export function printChain(chainId: number) {
-  if(!(chainId in CHAIN_ID_TO_NAME)) throw Error(`Invalid Chain ID: ${chainId}`);
+  if (!(chainId in CHAIN_ID_TO_NAME))
+    throw Error(`Invalid Chain ID: ${chainId}`);
   return `${CHAIN_ID_TO_NAME[chainId as ChainId]} (Chain ${chainId})`;
 }
 
-export function getDefaultProvider(network: Network, chain: ChainName, ci?: boolean) {
+export function getDefaultProvider(
+  network: Network,
+  chain: ChainName,
+  ci?: boolean
+) {
   let rpc: string | undefined = "";
-  if(ci) {
-    if(chain == "ethereum") rpc = "http://eth-devnet:8545";
-    else if(chain == "bsc") rpc = "http://eth-devnet2:8545";
-    else throw Error(`This chain isn't in CI for relayers: ${chain}`)
+  if (ci) {
+    if (chain == "ethereum") rpc = "http://eth-devnet:8545";
+    else if (chain == "bsc") rpc = "http://eth-devnet2:8545";
+    else throw Error(`This chain isn't in CI for relayers: ${chain}`);
   } else {
     rpc = RPCS_BY_CHAIN[network][chain];
   }
-  if(!rpc) {
+  if (!rpc) {
     throw Error(`No default RPC for chain ${chain} or network ${network}`);
   }
   return new ethers.providers.StaticJsonRpcProvider(rpc);
@@ -100,7 +138,7 @@ export async function getWormholeRelayerInfoBySourceSequence(
   blockStartNumber: ethers.providers.BlockTag,
   blockEndNumber: ethers.providers.BlockTag,
   targetWormholeRelayerAddress: string
-): Promise<{chain: ChainName, events: DeliveryTargetInfo[]}> {
+): Promise<{ chain: ChainName; events: DeliveryTargetInfo[] }> {
   const deliveryEvents = await getWormholeRelayerDeliveryEventsBySourceSequence(
     environment,
     targetChain,
@@ -116,9 +154,9 @@ export async function getWormholeRelayerInfoBySourceSequence(
     try {
       const blockStart = await targetChainProvider.getBlock(blockStartNumber);
       const blockEnd = await targetChainProvider.getBlock(blockEndNumber);
-      status = `Delivery didn't happen on ${targetChain} within blocks ${blockStart.number} to ${
-        blockEnd.number
-      } (within times ${new Date(
+      status = `Delivery didn't happen on ${targetChain} within blocks ${
+        blockStart.number
+      } to ${blockEnd.number} (within times ${new Date(
         blockStart.timestamp * 1000
       ).toString()} to ${new Date(blockEnd.timestamp * 1000).toString()})`;
     } catch (e) {}
@@ -129,12 +167,12 @@ export async function getWormholeRelayerInfoBySourceSequence(
       sourceChain: sourceChain,
       sourceVaaSequence,
       gasUsed: BigNumber.from(0),
-      refundStatus: RefundStatus.RefundFail
+      refundStatus: RefundStatus.RefundFail,
     });
   }
   const targetChainStatus = {
     chain: targetChain,
-    events: deliveryEvents
+    events: deliveryEvents,
   };
 
   return targetChainStatus;
@@ -151,7 +189,7 @@ export async function getWormholeRelayerDeliveryEventsBySourceSequence(
   targetWormholeRelayerAddress: string
 ): Promise<DeliveryTargetInfo[]> {
   const sourceChainId = CHAINS[sourceChain];
-  if(!sourceChainId) throw Error(`Invalid source chain: ${sourceChain}`)
+  if (!sourceChainId) throw Error(`Invalid source chain: ${sourceChain}`);
   const wormholeRelayer = getWormholeRelayer(
     targetChain,
     environment,
@@ -165,13 +203,23 @@ export async function getWormholeRelayerDeliveryEventsBySourceSequence(
     sourceVaaSequence
   );
 
-  const deliveryEventsPreFilter: DeliveryEvent[] = await wormholeRelayer.queryFilter(
-    deliveryEvents,
-    blockStartNumber,
-    blockEndNumber
-  );
+  const deliveryEventsPreFilter: DeliveryEvent[] =
+    await wormholeRelayer.queryFilter(
+      deliveryEvents,
+      blockStartNumber,
+      blockEndNumber
+    );
 
-  const isValid: boolean[] = await Promise.all(deliveryEventsPreFilter.map((deliveryEvent) => areSignaturesValid(deliveryEvent.getTransaction(), targetChain, targetChainProvider, environment)));
+  const isValid: boolean[] = await Promise.all(
+    deliveryEventsPreFilter.map((deliveryEvent) =>
+      areSignaturesValid(
+        deliveryEvent.getTransaction(),
+        targetChain,
+        targetChainProvider,
+        environment
+      )
+    )
+  );
 
   // There is a max limit on RPCs sometimes for how many blocks to query
   return await transformDeliveryEvents(
@@ -180,22 +228,36 @@ export async function getWormholeRelayerDeliveryEventsBySourceSequence(
   );
 }
 
-async function areSignaturesValid(transaction: Promise<ethers.Transaction>, targetChain: ChainName, targetChainProvider: ethers.providers.Provider, environment: Network) {
+async function areSignaturesValid(
+  transaction: Promise<ethers.Transaction>,
+  targetChain: ChainName,
+  targetChainProvider: ethers.providers.Provider,
+  environment: Network
+) {
   const coreAddress = CONTRACTS[environment][targetChain].core;
-  if(!coreAddress) throw Error(`No Wormhole Address for chain ${targetChain}, network ${environment}`);
+  if (!coreAddress)
+    throw Error(
+      `No Wormhole Address for chain ${targetChain}, network ${environment}`
+    );
 
-  const wormhole = Implementation__factory.connect(coreAddress, targetChainProvider);
-  const decodedData = IWormholeRelayerDelivery__factory.createInterface().parseTransaction(await transaction);
+  const wormhole = Implementation__factory.connect(
+    coreAddress,
+    targetChainProvider
+  );
+  const decodedData =
+    IWormholeRelayerDelivery__factory.createInterface().parseTransaction(
+      await transaction
+    );
 
   const vaaIsValid = async (vaa: ethers.utils.BytesLike): Promise<boolean> => {
-    const [,result,reason] = await wormhole.parseAndVerifyVM(vaa);
-    if(!result) console.log(`Invalid vaa! Reason: ${reason}`);
+    const [, result, reason] = await wormhole.parseAndVerifyVM(vaa);
+    if (!result) console.log(`Invalid vaa! Reason: ${reason}`);
     return result;
-  }
+  };
 
   const vaas = decodedData.args[0];
-  for(let i=0; i<vaas.length; i++) {
-    if(!(await vaaIsValid(vaas[i]))) {
+  for (let i = 0; i < vaas.length; i++) {
+    if (!(await vaaIsValid(vaas[i]))) {
       return false;
     }
   }
@@ -222,12 +284,11 @@ async function transformDeliveryEvents(
   events: DeliveryEvent[],
   targetProvider: ethers.providers.Provider
 ): Promise<DeliveryTargetInfo[]> {
-
-
   return Promise.all(
     events.map(async (x) => {
       const status = deliveryStatus(x.args[4]);
-      if(!isChain(x.args[1])) throw Error(`Invalid source chain id: ${x.args[1]}`);
+      if (!isChain(x.args[1]))
+        throw Error(`Invalid source chain id: ${x.args[1]}`);
       const sourceChain = CHAIN_ID_TO_NAME[x.args[1] as ChainId];
       return {
         status,
@@ -237,8 +298,19 @@ async function transformDeliveryEvents(
         sourceChain,
         gasUsed: BigNumber.from(x.args[5]),
         refundStatus: x.args[6],
-        revertString: (status == DeliveryStatus.ReceiverFailure) ? x.args[7] : (status == DeliveryStatus.ForwardRequestFailure ? parseForwardFailureError(Buffer.from(x.args[7].substring(2), "hex")): undefined),
-        overridesInfo: (Buffer.from(x.args[8].substring(2), "hex").length > 0) && parseOverrideInfoFromDeliveryEvent(Buffer.from(x.args[8].substring(2), "hex"))
+        revertString:
+          status == DeliveryStatus.ReceiverFailure
+            ? x.args[7]
+            : status == DeliveryStatus.ForwardRequestFailure
+            ? parseForwardFailureError(
+                Buffer.from(x.args[7].substring(2), "hex")
+              )
+            : undefined,
+        overridesInfo:
+          Buffer.from(x.args[8].substring(2), "hex").length > 0 &&
+          parseOverrideInfoFromDeliveryEvent(
+            Buffer.from(x.args[8].substring(2), "hex")
+          ),
       };
     })
   );
@@ -288,9 +360,7 @@ export function getWormholeRelayerLog(
   }
 }
 
-export function vaaKeyToVaaKeyStruct(
-  vaaKey: VaaKey
-): VaaKeyStruct {
+export function vaaKeyToVaaKeyStruct(vaaKey: VaaKey): VaaKeyStruct {
   return {
     chainId: vaaKey.chainId || 0,
     emitterAddress:
